# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Waves API
  version: 1.0.0
  summary: Read and write a Waves shared ledger — groups, expenses, settlements and balances.
  description: |
    The Waves API lets a program act for a Waves user: read the groups they are
    in, add and edit expenses, record payments, and read what everybody is owed.
    Every request runs as that person with their row-level security intact, so a
    token can never reach a group its owner could not reach from the app. A scope
    only narrows what the token may do; it never widens what the person can see.

    ## Authenticating

    There are two ways in, and they produce the same kind of credential.

    **A personal access token** is a developer acting on their own account. It is
    created in the Waves developer console, carries a fixed scope list, and is
    sent as `Authorization: Bearer wavs_pat_…`. Use this when the program and the
    account belong to the same person.

    **OAuth 2.1 authorization code with PKCE** is somebody else's program acting
    for a Waves user who approved it. Send the browser to `/oauth/authorize`,
    exchange the returned code at `/oauth/token`, and use the resulting
    `wavs_at_…` access token in the same `Authorization` header. `S256` is the
    only code challenge method accepted, and a refresh token is only issued when
    `offline_access` was granted. Refresh tokens rotate: each one is single-use,
    and replaying a spent one disconnects the application.

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /oauth/authorize:
    get:
      tags: [OAuth]
      operationId: authorize
      summary: Start the authorization code flow
      description: |
        Checks the *shape* of the request and redirects the browser to this
        deployment's own consent screen, carrying the parameters along. It never
        redirects to a caller-supplied address and never touches the database,
        which is why it cannot be used to probe whether a client id exists —
        whether the client and the redirect are real is decided later, under the
        signed-in user's session.

        Errors are returned as an RFC 6749 JSON body rather than bounced to
        `redirect_uri`: a validator that redirected errors to an unverified
        address would be an open redirect wearing an OAuth costume.
      security: []
      parameters:
        - name: response_type
          in: query
          required: true
          description: Must be `code`.
          schema:
            type: string
            enum: [code]
        - name: client_id
          in: query
          required: true
          description: The application's public client id.
          schema:
            type: string
            examples: ['wavs_app_5f2c1d9a4b7e42f1a0c6d3e8b91f7a24']
        - name: redirect_uri
          in: query
          required: true
          description: >-
            Where to deliver the code. Must be one of the addresses registered
            for this application — checked later, when the user approves.
          schema:
            type: string
            format: uri
        - name: scope
          in: query
          required: true
          description: >-
            Space-separated. Every entry must be a known scope; one unknown
            entry fails the whole request rather than being silently dropped.
          schema:
            type: string
            examples: ['identity.read groups.read expenses.write offline_access']
        - name: code_challenge
          in: query
          required: true
          description: The unpadded base64url SHA-256 of your verifier — 43 characters.
          schema:
            type: string
            pattern: '^[A-Za-z0-9_-]{43}$'
        - name: code_challenge_method
          in: query
          required: true
          description: '`S256` only. `plain` is a challenge whose answer is in the same request.'
          schema:
            type: string
            enum: [S256]
        - name: state
          in: query
          required: false
          description: Returned unchanged on the redirect back.
          schema:
            type: string
      responses:
        '302':
          description: >-
            Redirect to this deployment's consent screen. The user approves
            there, and the screen calls `POST /oauth/authorize` with their own
            session.
          headers:
            Location:
              description: The consent screen, with the request's parameters attached.
              schema:
                type: string
                format: uri
        '400':
          description: >-
            The request is malformed, or `scope` names something this deployment
            does not know.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthError' }
        '500':
          description: >-
            This deployment has no consent screen configured
            (`WAVES_API_WEB_URL` is unset).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthError' }
    post:
      tags: [OAuth]
      operationId: approveAuthorization
      summary: Record the user's approval and mint a code
      description: >-
        Called by the consent screen, authenticated by the *user's own Waves
        session* rather than a client credential — the decision being recorded
        is theirs. The redirect and the scopes are re-checked against the
        application's registration here, because the browser between the two
        screens is the attacker's ground and nothing it carries is believed
        twice. Returns the address to send the browser to, rather than
        redirecting, so the screen can control the navigation.


        Because it carries a session, this endpoint is on the console's CORS
        surface, not the public one: a browser can only call it from an origin
        in the deployment's `WAVES_API_ALLOWED_ORIGINS` allowlist, and a
        deployment that has not set that variable refuses browser calls
        entirely. The `GET` beside it carries no credential and stays public.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveAuthorizationRequest'
      responses:
        '200':
          description: Where to send the browser, with the code attached.
          content:
            application/json:
              schema:
                type: object
                required: [redirect_to]
                properties:
                  redirect_to:
                    type: string
                    format: uri
                    description: >-
                      The application's registered redirect with `code` and, if
                      you sent one, `state` appended.
        '400':
          description: Malformed, or the redirect or scopes are not registered for this client.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthError' }
        '401':
          description: No Waves session, or one the authorization server would not vouch for.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthError' }
        '500':
          description: Something went wrong on our side.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthError' }
...
  /developer/apps:
    get:
      tags: [Developer console]
      operationId: listApps
      summary: Applications this person has registered
      description: Newest first. Client secrets are never returned. Not paginated.
      security:
        - sessionAuth: []
      responses:
        '200':
          description: The applications.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppList'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [Developer console]
      operationId: registerApp
      summary: Register an OAuth application
      description: >-
        A confidential client is one that can actually keep a secret — a server.
        A mobile or single-page app cannot, and a secret published in an app
        bundle is not a secret, so send `confidential: false` for those and rely
        on PKCE. A `client_secret` is returned only for a confidential client,
        and only here. An account that already holds as many applications as it
        is allowed gets a 422 (`APP_LIMIT`).
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAppRequest'
      responses:
        '201':
          description: The registration, with the client secret shown once.
          headers:
            Cache-Control:
              description: Always `no-store`.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisteredApp'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /developer/apps/{appId}:
    parameters:
      - $ref: '#/components/parameters/AppId'
    patch:
      tags: [Developer console]
      operationId: updateApp
      summary: Change an application's registration
      description: >-
        Every field is optional; an omitted one is left alone. The client secret
        is never rotated as a side effect of another edit — that is its own call.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAppRequest'
      responses:
        '200':
          description: The application as it now stands.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/App'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      tags: [Developer console]
      operationId: deleteApp
      summary: Delete an application
      description: Owner-only. Every token issued to it stops working.
      security:
        - sessionAuth: []
      responses:
        '204':
          description: Deleted.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: wavs_pat_… or wavs_at_…
      description: |
        A Waves API token in `Authorization: Bearer …`. Two kinds are accepted
        here and they behave identically once presented:

        - `wavs_pat_…` — a personal access token, created in the developer
          console, acting for the developer's own account.
        - `wavs_at_…` — an OAuth access token, acting for a user who approved an
          application.

        A refresh token (`wavs_rt_…`) is a credential for `/oauth/token` and
        nothing else; presenting one here is refused with `invalid_token`.

        The token string carries no scope list — the row does — so the scopes
        each operation lists apply to both kinds, and narrowing an
        application's scopes takes effect on its next request.

        A missing token is `401 unauthorized` with
        `WWW-Authenticate: Bearer realm="waves"`. A token that is malformed,
        expired, revoked, or belongs to a disabled application is `401
        invalid_token` with
        `WWW-Authenticate: Bearer realm="waves", error="invalid_token"`. A live
        token that lacks the scope the route needs is `403 insufficient_scope`.
    oauth2:
      type: oauth2
      description: >-
        OAuth 2.1 authorization code with PKCE (`S256` only). The URLs below are
        relative to this deployment's own base URL; the authoritative copies are
        in `/.well-known/oauth-authorization-server`, which derives them from
        the request. Ask for `offline_access` if you need a refresh token.
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          refreshUrl: /oauth/token
          scopes:
            identity.read: See your name, avatar, country, language, default currency and payment
...
  headers:
    XRequestId:
      description: >-
        This request's id. Present on every response, including errors, and
        repeated in the error body as `request_id`.
      schema:
        type: string
        format: uuid
    RateLimitLimit:
      description: >-
        Requests allowed in the current window — the lower of the per-token and
        per-account budgets. Present on every response that got as far as
        authenticating.
      schema:
        type: integer
    RateLimitRemaining:
      description: Requests left in the current window.
      schema:
        type: integer
    RateLimitReset:
      description: >-
        Seconds from now until the window resets — delta-seconds, the same unit
        as `Retry-After` on the 429 that follows, and what the IETF
        RateLimit-header drafts specify. Never negative.
      schema:
        type: integer
        minimum: 0
...
    WWWAuthenticate:
      description: >-
        `Bearer realm="waves"` when no token was sent, and
        `Bearer realm="waves", error="invalid_token"` when one was sent and is
        not usable.
      schema:
        type: string

  # ───────────────────────────────────────────────── shared parameters ──
  parameters:
...
    AppId:
      name: appId
      in: path
      required: true
      description: The application's id — not its `client_id`.
      schema:
        type: string
        format: uuid

  # ────────────────────────────────────────────────── shared responses ──
  responses:
    BadRequest:
      description: >-
        `invalid_request` — the body, a query parameter or a header is not
        something this endpoint can act on. The message names the field.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: >-
        `unauthorized` when no credential was sent, or `invalid_token` when one
        was sent and is malformed, expired, revoked, of the wrong kind, or
        belongs to a disabled application.
      headers:
        WWW-Authenticate: { $ref: '#/components/headers/WWWAuthenticate' }
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: >-
        `insufficient_scope` when the token is live but was not granted the
        scope this route needs, or `forbidden` when the account itself may not
        do this — not a member of the group, not a party to the settlement, not
        the owner, or a guest whose trial has ended.
      headers:
        WWW-Authenticate: { $ref: '#/components/headers/WWWAuthenticate' }
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
        RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
        RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
        RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: >-
        `not_found`. Note that a row the caller is not allowed to see is absent
        rather than forbidden, so this also covers "it exists, but not for you".
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
        RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
        RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
        RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: >-
        `conflict` — the write collided with the state it was based on. A stale
        `if_revision` or `base_version_no`, shares that do not reconcile with
        the amount, or an id that is already taken.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
        RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
        RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
        RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unprocessable:
      description: >-
        `unprocessable` — the request is well formed and you are allowed to make
        it, but the ledger's own rules say no. A limit (the applications or live
        tokens an account may have), a group that is not settled, the last admin
        being demoted, a restore window that has closed, or a database CHECK
        constraint.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    InternalError:
      description: >-
        `internal` when something went wrong on our side, or `misconfigured`
        when the deployment is missing configuration it needs (check `/health`).
        Neither says anything about the cause; quote `request_id`.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }

  # ──────────────────────────────────────────────────────────── schemas ──
  schemas:
...
    ScopeName:
      type: string
      enum:
        - identity.read
        - identity.write
        - groups.read
        - groups.write
        - expenses.read
        - expenses.write
        - settlements.read
        - settlements.write
        - friends.read
        - categories.read
        - categories.write
        - offline_access
      description: >-
        One of the scopes this deployment knows. An unknown entry fails the
        whole request rather than being dropped, because silently granting less
        than was asked for is how a client ends up mysteriously unable to do the
        thing it thought it had permission for.
...
    Error:
      type: object
      description: >-
        Every refusal under `/v1` and `/developer` has this shape. It is nested
        under `error` so that it can never collide with a resource that happens
        to have a `message`, and so a client can branch on `"error" in body`
        without knowing which endpoint it called. Nothing internal — a database
        message, a stack, a hostname — ever reaches this body.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              enum:
                - invalid_request
                - unauthorized
                - invalid_token
                - insufficient_scope
                - forbidden
                - not_found
                - conflict
                - unprocessable
                - rate_limited
                - misconfigured
                - internal
              description: |
                A token to match on, rather than a sentence to parse. The HTTP
                status each one carries:

                | code | status |
                | --- | --- |
                | `invalid_request` | 400 |
                | `unauthorized` | 401 |
                | `invalid_token` | 401 |
                | `insufficient_scope` | 403 |
                | `forbidden` | 403 |
                | `not_found` | 404 |
                | `conflict` | 409 |
                | `unprocessable` | 422 |
                | `rate_limited` | 429 |
                | `misconfigured` | 500 |
                | `internal` | 500 |
            message:
              type: string
              description: >-
                A sentence written for a person. Safe to show, and safe to
                ignore — branch on `code`, not on this.
            request_id:
              type: string
              format: uuid
              description: The same value as the `X-Request-Id` header. Quote it when reporting a problem.
            details:
              type: object
              additionalProperties: true
              description: >-
                Present only when naming the offending thing is what makes the
                error useful. `{"editable": [...]}` on a patch with nothing
                patchable in it, `{"field": "default_currency"}` on a value of
                the wrong shape, `{"parameter": "include_archived"}` on an
                unreadable query flag, `{"header": "Idempotency-Key"}` on a key
                of the wrong length, or `{"scopes": [...]}` listing the
                catalogue when an unknown scope was named.
      examples:
        - error:
            code: insufficient_scope
            message: this token was not granted that
            request_id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
    OAuthError:
      type: object
      description: >-
        RFC 6749's error shape, used only by the `/oauth/*` endpoints. Note that
        `error` is a **string** here, not the object the rest of the API
        returns — this is the one place the service deliberately deviates from
        its own convention, because it is what every OAuth client library
        parses.
      required: [error, error_description]
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
            - server_error
        error_description:
          type: string
      examples:
        - error: invalid_grant
          error_description: that is not a code we issued

    # ── resources ──
...
    ApproveAuthorizationRequest:
      type: object
      description: >-
        The same parameters `GET /oauth/authorize` takes, as a JSON body, posted
        by the consent screen once the person has approved.
      required:
        [response_type, client_id, redirect_uri, scope, code_challenge, code_challenge_method]
      properties:
        response_type:
          type: string
          enum: [code]
        client_id:
          type: string
        redirect_uri:
          type: string
          format: uri
        scope:
          type: string
          description: Space-separated.
        code_challenge:
          type: string
          pattern: '^[A-Za-z0-9_-]{43}$'
        code_challenge_method:
          type: string
          enum: [S256]
        state:
          type: string
...
    App:
      type: object
      description: A registered OAuth application, as its owner sees it.
      required:
        [
          id,
          name,
          description,
          website_url,
          client_id,
          redirect_uris,
          scopes,
          disabled_at,
          created_at,
        ]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        website_url:
          type: [string, 'null']
          format: uri
        client_id:
          type: string
          examples: ['wavs_app_5f2c1d9a4b7e42f1a0c6d3e8b91f7a24']
        redirect_uris:
          type: array
          items:
            type: string
            format: uri
        scopes:
          type: array
          description: The most this application may ever ask a user for.
          items:
            $ref: '#/components/schemas/ScopeName'
        disabled_at:
          type: [string, 'null']
          format: date-time
          description: While set, every token issued to this application is refused.
        created_at:
          type: string
          format: date-time
    AppList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/App'
    RegisterAppRequest:
      type: object
      required: [name, scopes, redirect_uris]
      properties:
        name:
          type: string
          minLength: 1
        scopes:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ScopeName'
        redirect_uris:
          type: array
          minItems: 1
          maxItems: 10
          description: Between 1 and 10 addresses. Every one must be a non-empty string.
          items:
            type: string
            format: uri
        description:
          type: string
          default: ''
        website_url:
          type: [string, 'null']
          format: uri
        confidential:
          type: boolean
          default: true
          description: >-
            True mints a client secret. Send false for a mobile or single-page
            app, which cannot keep one, and rely on PKCE instead.
      examples:
        - name: Trip Budget Bot
          description: Posts a nightly summary of a trip group to Slack.
          website_url: https://tripbudget.example.com
          redirect_uris: ['https://tripbudget.example.com/oauth/callback']
          scopes: [identity.read, groups.read, expenses.read, offline_access]
          confidential: true
    RegisteredApp:
      type: object
      required: [id, name, client_id, redirect_uris, scopes]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        client_id:
          type: string
        redirect_uris:
          type: array
          items:
            type: string
            format: uri
        scopes:
          type: array
          items:
            $ref: '#/components/schemas/ScopeName'
        client_secret:
          type: string
          description: >-
            Present only for a confidential client, and only here. Rotating it
            later is a separate, deliberate call.
          examples: ['wavs_cs_Zm9vYmFyYmF6cXV4Y29ycmdlZ3JhdWx0']
    UpdateAppRequest:
      type: object
      description: Every field optional; an omitted one is left alone.
      properties:
        name:
          type: string
        description:
          type: string
        website_url:
          type: string
          format: uri
        redirect_uris:
          type: array
          minItems: 1
          maxItems: 10
          items:
            type: string
            format: uri
        scopes:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ScopeName'
        disabled:
          type: boolean
          description: True refuses every token issued to this application until it is set back.
      examples:
        - redirect_uris:
            ['https://tripbudget.example.com/oauth/callback', 'http://localhost:5173/callback']
...